requirements.txt
# Django
asgiref==3.5.0
Django==4.0.4
sqlparse==0.4.2
tzdata==2022.1
# DRF
djangorestframework==3.13.1
pytz==2022.1
# Autopep8
autopep8==1.6.0
pycodestyle==2.8.0
toml==0.10.2
# cors
django-cors-headers==3.11.0

# pylint
astroid==2.11.3
colorama==0.4.4
dill==0.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.13.7
tomli==2.0.1
wrapt==1.14.0